Movimiento del Guardian Definido

+++ models/Guardian.java
@@ -4,7 +4,7 @@ import lupin.models.abstracts.*;
 
 public class Guardian extends Enemigo {
 
-    public Guardian(int x,int y, int velocidad){
+    public Guardian(double x,double y, int velocidad){
         super(x,y, velocidad);
     }
 }
